Update README.md
This projects is motivated by a personal real-world problem: When we moved into a new flat, I damaged a small plastic part (picture to the right) of an already installed wall lamp. The broken plastic leg caused the lid to fall off. I searched the Internet for spare parts – no luck. According to the manufacturer, the only option is to repurchase the same lamp. Not very sustainable. 

I found a website that claimed to provide a 3D model of the plastic part that could be printed using a 3D printer – but the website was not working and thus not providing the file. Printing simple plastic spare parts would have solved the issue. Problem: Like most people, I do not have a background in 3D printing and do not have the time to learn how to construct the corresponding models. A less sophisticated solution is required, requiring neither special hardware nor special knowledge.

The goal of this research is to develop a method to conveniently and effortlessly print 3D spare parts for everyday life repairs (such as the wall lamp above). The overall thesis is separated into three sub-tasks: 

1.	Taking the broken component or another (not broken) spare part (if available) and capturing its dimensions using a suitable digitization method (e.g., taking pictures, videos, etc.). 

2.	Processing the captured information, producing a 3D printable model of the component. 

3.	Printing the component. 
The thesis focuses on simple geometric components such as the broken plastic part in the picture above. Subsequent research may tackle more complex structures.

With the results of the research, we aim to make the provision of 3D printed spare parts more accessible to everyone, reduce the complexity of the process and automate as many steps as possible along the way. This way, the research also aims to generate tangible contributions to a more sustainable future.